1. small modify;

+++ compute/src/bean/CallTimeBean.java
@@ -5,7 +5,7 @@ package bean;
  * All right reserved.
  */
 public class CallTimeBean {
-    public long timeSlot1 = 0,  // 0~3
+    public int  timeSlot1 = 0,  // 0~3
                 timeSlot2 = 0,  // 3~6
                 timeSlot3 = 0,  // 6~9
                 timeSlot4 = 0,  // 9~12
